fix: readd modified override-types
